requirements.txt
pandas
gspread
google-auth
supabase
python-dotenv
